Fixed autofocus error
In my browser console i had an autofocus error related to my sign up page. Every input field had the autofocus but it is only alowed to have in on one input field that is focusing when the page is loaded. This makes sure the user can start typing on the field where the autofocus is placed